fix: set output to os.Stdout and os.Stderr
Previously, the Execute function would output the initial stderr
or stdout value when executing the command. However, further output
generated by the command (such as program logs) is no longer
available after wait-for-it exits. This update replaces the locally
defined buffers with os.Stdout and os.Stderr so that output
continues to be captured.